backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.1.2
python-multipart==0.0.6
pandas==2.1.4
openpyxl==3.1.2
xlrd==2.0.1
pyxlsb==1.0.10
openai==1.3.7
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1